Controlling opacity of HTML object with JS
Hi there, I am trying to learn a bit more about website development. In particular at this point in time I'm trying to achieve controlling the opacity of two objects, making the opacity oscilate between 0% to 100% continuosly. I tried doing this with keyframes in the css file but it doesn't loop as smoothly as I would like it. 

Could any of you talented coders help a newbie find some direction on why my code is not compiling?

Thank you so much!